Update Servlets. - AnnotationServlet.java replaces {write,remove}Annotations*.java - both POST and DELETE requests via '/annotation' urlpath - Servlet Specification v2.5 -> 3.0 - character encoding set to UTF-8
