+++ pom.xml
@@ -8,56 +8,44 @@
   <name>semanticAnnotator</name>
   <url>http://maven.apache.org</url>
   <dependencies>
-    <!--   <dependency>
-    <groupId>com.thetransactioncompany</groupId>
-    <artifactId>cors-filter</artifactId>
-    <version>1.3.2</version>
- </dependency> -->
-    <!-- Junit -->
     <dependency>
       <groupId>junit</groupId>
       <artifactId>junit</artifactId>
       <version>3.8.1</version>
       <scope>test</scope>
     </dependency>
-    <!-- RDF4j -->
     <dependency>
       <groupId>org.eclipse.rdf4j</groupId>
       <artifactId>rdf4j-runtime</artifactId>
       <version>3.5.0</version>
       <type>pom</type>
     </dependency>
-    <!-- logback -->
+    <dependency>
+      <groupId>org.eclipse.rdf4j</groupId>
+      <artifactId>rdf4j-repository-sparql</artifactId>
+      <version>3.5.0</version>
+    </dependency>
     <dependency>
       <groupId>ch.qos.logback</groupId>
       <artifactId>logback-classic</artifactId>
       <version>1.0.13</version>
     </dependency>
-    <!-- javax -->
     <dependency>
       <groupId>javax.servlet</groupId>
-      <artifactId>servlet-api</artifactId>
-      <version>3.0-alpha-1</version>
+      <artifactId>javax.servlet-api</artifactId>
+      <version>3.1.0</version>
+      <scope>provided</scope>
     </dependency>
-    <!-- Tiles -->
-    <dependency>
-      <groupId>org.apache.tiles</groupId>
-      <artifactId>tiles-extras</artifactId>
-      <version>3.0.4</version>
-    </dependency>
-    <!-- JSTL -->
     <dependency>
       <groupId>javax.servlet</groupId>
       <artifactId>jstl</artifactId>
       <version>1.2</version>
     </dependency>
-    <!-- mySQL
     <dependency>
-      <groupId>mysql</groupId>
-      <artifactId>mysql-connector-java</artifactId>
-      <version>5.1.6</version>
-    </dependency> -->
-    <!-- JSON.simple -->
+      <groupId>org.apache.tiles</groupId>
+      <artifactId>tiles-extras</artifactId>
+      <version>3.0.4</version>
+    </dependency>
     <dependency>
       <groupId>com.googlecode.json-simple</groupId>
       <artifactId>json-simple</artifactId>
@@ -68,19 +56,11 @@
       <artifactId>httpcore</artifactId>
       <version>4.4.5</version>
     </dependency>
-    <!-- SPARQL ENDPOINTS -->
-    <dependency>
-      <groupId>org.eclipse.rdf4j</groupId>
-      <artifactId>rdf4j-repository-sparql</artifactId>
-      <version>3.5.0</version>
-    </dependency>
-    <!-- update jsonld -->
     <dependency>
       <groupId>com.github.jsonld-java</groupId>
       <artifactId>jsonld-java</artifactId>
       <version>0.10.0</version>
     </dependency>
-    <!-- json -->
     <dependency>
       <groupId>org.json</groupId>
       <artifactId>json</artifactId>
