+++ src/main/webapp/WEB-INF/web.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<web-app xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://java.sun.com/xml/ns/javaee" xmlns:web="http://java.sun.com/xml/ns/javaee" xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-app_2_5.xsd" id="semanticAnnotator" version="2.5">
+<web-app xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://java.sun.com/xml/ns/javaee" xmlns:web="http://java.sun.com/xml/ns/javaee" xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-app_3_0.xsd" id="semanticAnnotator" version="3.0">
   <display-name>semanticAnnotator</display-name>
   <servlet>
     <servlet-name>Tiles Dispatch Servlet</servlet-name>
@@ -11,9 +11,7 @@
   </servlet-mapping>
   <context-param>
     <param-name>org.apache.tiles.definition.DefinitionsFactory.DEFINITIONS_CONFIG</param-name>
-    <param-value>
-            /WEB-INF/tiles.xml
-        </param-value>
+    <param-value>/WEB-INF/tiles.xml</param-value>
   </context-param>
   <listener>
     <listener-class>org.apache.tiles.extras.complete.CompleteAutoloadTilesListener</listener-class>
@@ -21,6 +19,14 @@
   <welcome-file-list>
     <welcome-file>index.jsp</welcome-file>
   </welcome-file-list>
+  <filter>
+    <filter-name>Set Character Encoding</filter-name>
+    <filter-class>org.apache.catalina.filters.SetCharacterEncodingFilter</filter-class>
+    <init-param>
+      <param-name>encoding</param-name>
+      <param-value>UTF-8</param-value>
+    </init-param>
+  </filter>
   <filter>
     <filter-name>CorsFilter</filter-name>
     <filter-class>org.apache.catalina.filters.CorsFilter</filter-class>
@@ -53,24 +59,6 @@
     <filter-name>CorsFilter</filter-name>
     <url-pattern>/*</url-pattern>
   </filter-mapping>
-  <servlet>
-    <servlet-name>FileServlet</servlet-name>
-    <servlet-class>localdomain.localhost.servlets.FileServlet</servlet-class>
-  </servlet>
-  <servlet-mapping>
-    <servlet-name>FileServlet</servlet-name>
-    <url-pattern>/files/*</url-pattern>
-  </servlet-mapping>
-  <servlet>
-    <description/>
-    <display-name>WriteAnnotations</display-name>
-    <servlet-name>WriteAnnotations</servlet-name>
-    <servlet-class>localdomain.localhost.servlets.WriteAnnotations</servlet-class>
-  </servlet>
-  <servlet-mapping>
-    <servlet-name>WriteAnnotations</servlet-name>
-    <url-pattern>/WriteAnnotations</url-pattern>
-  </servlet-mapping>
   <security-role>
     <role-name>manager-gui</role-name>
   </security-role>
@@ -86,24 +74,4 @@
   <login-config>
     <auth-method>BASIC</auth-method>
   </login-config>
-  <servlet>
-    <description/>
-    <display-name>writeAnnotationsToRDF</display-name>
-    <servlet-name>writeAnnotationsToRDF</servlet-name>
-    <servlet-class>localdomain.localhost.servlets.writeAnnotationsToRDF</servlet-class>
-  </servlet>
-  <servlet-mapping>
-    <servlet-name>writeAnnotationsToRDF</servlet-name>
-    <url-pattern>/writeAnnotationsToRDF</url-pattern>
-  </servlet-mapping>
-  <servlet>
-    <description/>
-    <display-name>removeAnnotationsFromRDF</display-name>
-    <servlet-name>removeAnnotationsFromRDF</servlet-name>
-    <servlet-class>localdomain.localhost.servlets.removeAnnotationsFromRDF</servlet-class>
-  </servlet>
-  <servlet-mapping>
-    <servlet-name>removeAnnotationsFromRDF</servlet-name>
-    <url-pattern>/removeAnnotationsFromRDF</url-pattern>
-  </servlet-mapping>
 </web-app>
