--- test/openseadragon.html
@@ -1,65 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta http-equiv="content-type" content="text/html; charset=UTF-8">
-    <title>Annotorious | OpenSeadragon Plugin</title>
-    <link rel="stylesheet" href="annotorious.min.css">
-    <script type="text/javascript" src="openseadragon.min.js"></script>
-    <script type="text/javascript" src="openseadragon-annotorious.min.js"></script>
-  </head>
-  <body>
-  <div id="openseadragon" style="width: 640px; height: 480px; position: relative;"></div>
-  <script>
-  var sampleAnnotation = { 
-    "@context": "http://www.w3.org/ns/anno.jsonld",
-    "id": "#07475897-d2eb-4dce-aa12-ecb50771c734",
-    "type": "Annotation",
-    "body": [{
-      "type": "TextualBody",
-      "value": "Annotation"
-    }],
-    "target": {
-      "selector": {
-        "type": "FragmentSelector",
-        "conformsTo": "http://www.w3.org/TR/media-frags/",
-        "value": "xywh=200,200,10,100"
-      }
-    }
-  };
-
-  window.onload = function() {
-    var viewer = OpenSeadragon({
-      id: "openseadragon",
-      //prefixUrl: "/js/openseadragon/images/",
-      preserveViewport: true,
-      tileSources: [
-        "http://localhost:8182/iiif/2/900c341c1c10fff7:MMNAT01_PM_NNM001001033_001/info.json"
-      ],
-    });
-    // Initialize the Annotorious plugin
-    var anno = OpenSeadragon.Annotorious(viewer, {
-      locale: "auto"
-    });
-
-    anno.setAuthInfo({
-      id: 'http://example.com/arni',
-      displayName: 'arni'
-    });
-
-    // Load annotations in W3C Web Annotation format
-    anno.loadAnnotations(sampleAnnotation);
-    
-    // Attach listeners to handle annotation events
-    anno.on("createAnnotation", function(a) {
-      console.log("id=" + a.id);
-      console.log("creator=" + a.body[1].creator);
-      console.log("created=" + a.body[1].created);
-      console.log("targetSource=" + a.target.source);
-      console.log("targetSelector=" + a.target.selector.value);
-      console.log("bodyValue=" + a.body[0].value);
-      console.log("bodyPurpose=" + a.body[0].purpose);
-    });
-  }
-</script>
-</body>
-</html>
